requirements.txt
Flask==2.0.1
Flask-Bootstrap==3.3.7.1
Flask-Login==0.5.0
Flask-Moment==1.0.1
Flask-SQLAlchemy==2.5.1
Flask-WTF==0.15.1
Jinja2==3.0.1
#SQLAlchemy==1.4.20
Werkzeug==2.0.1
WTForms==2.3.3
WTForms-SQLAlchemy==0.2
email-validator==1.1.2
python-dotenv==0.18.0
# requirements for Heroku
#psycopg2==2.9.1
#gunicorn==20.1.0